Move test to shows module and move episode mock object to episode test class.
