+++ database/src/commonTest/kotlin/com/thomaskioko/tvmaniac/core/db/EpisodesCacheTest.kt
@@ -1,6 +1,5 @@
 package com.thomaskioko.tvmaniac.core.db
 
-import com.thomaskioko.tvmaniac.core.db.MockData.getEpisodeCacheList
 import com.thomaskioko.tvmaniac.database.test.BaseDatabaseTest
 import com.thomaskioko.tvmaniac.db.Id
 import io.kotest.matchers.shouldBe
@@ -44,4 +43,34 @@ internal class EpisodesCacheTest : BaseDatabaseTest() {
       image_url = image_url,
     )
   }
+
+  private fun getEpisodeCacheList() =
+    listOf(
+      Episode(
+        id = Id(2534997),
+        season_id = Id(114355),
+        show_id = Id(123232),
+        title = "Glorious Purpose",
+        overview =
+        "After stealing the Tesseract in Avengers: Endgame, Loki lands before the Time Variance Authority.",
+        vote_count = 42,
+        vote_average = 6.429,
+        runtime = 45,
+        episode_number = 1,
+        image_url = "/yDWJYRAwMNKbIYT8ZB33qy84uzO.jpg",
+      ),
+      Episode(
+        id = Id(2927202),
+        season_id = Id(114355),
+        show_id = Id(123232),
+        title = "The Variant",
+        overview =
+        "Mobius puts Loki to work, but not everyone at TVA is thrilled about the God of Mischief's presence.",
+        vote_count = 42,
+        vote_average = 6.429,
+        runtime = 45,
+        episode_number = 1,
+        image_url = "/yDWJYRAwMNKbIYT8ZB33qy84uzO.jpg",
+      ),
+    )
 }
